Consolidate optionsModule.ts tests.
